requirements.txt
beautifulsoup4==4.9.3
mysql-connector-python==8.0.26
requests==2.26.0
